requirements.txt
# Core requirements
numpy>=1.22.0
pandas>=1.4.0
scipy>=1.8.0
matplotlib>=3.5.0

# Image processing
scikit-image>=0.19.0
tifffile>=2021.11.2
pillow>=9.0.0

# Data analysis and machine learning
scikit-learn>=1.0.0

# Wavelet transform
dtcwt>=0.12.0

# Other utilities
setuptools>=60.0.0
wheel>=0.37.0 